+++ src/hash.h
@@ -17,7 +17,7 @@
 extern "C" {
 #endif
 
-#define CRC32_XINIT 0xFFFFFFFF		/* initial value */
+#define CRC32_XINIT 0xFFFFFFFF      /* initial value */
 #define CRC32_XOROT 0xFFFFFFFF		/* final xor value */
 
 typedef int (*HashFunc) (const void *key, const int key_len);
@@ -351,8 +351,8 @@ int simple_hash_ex(const void* key, const int key_len, \
 	const int init_value);
 
 int CRC32(const void *key, const int key_len);
-int CRC32_ex(const void *key, const int key_len, \
-	const int init_value);
+int64_t CRC32_ex(const void *key, const int key_len, \
+	const int64_t init_value);
 
 #define CRC32_FINAL(crc)  (crc ^ CRC32_XOROT)
